extract generic build functions to react-dev-utils
