+++ packages/react-dev-utils/getDifferenceLabel.js
@@ -1,0 +1,19 @@
+var chalk = require('chalk');
+var filesize = require('filesize');
+
+// Input: 1024, 2048
+// Output: "(+1 KB)"
+module.exports = function getDifferenceLabel(currentSize, previousSize) {
+  var FIFTY_KILOBYTES = 1024 * 50;
+  var difference = currentSize - previousSize;
+  var fileSize = !Number.isNaN(difference) ? filesize(difference) : 0;
+  if (difference >= FIFTY_KILOBYTES) {
+    return chalk.red('+' + fileSize);
+  } else if (difference < FIFTY_KILOBYTES && difference > 0) {
+    return chalk.yellow('+' + fileSize);
+  } else if (difference < 0) {
+    return chalk.green(fileSize);
+  } else {
+    return '';
+  }
+};
