+++ packages/react-dev-utils/paths/copyPublicFolder.js
@@ -1,0 +1,7 @@
+var fs = require('fs-extra');
+module.exports = paths => function copyPublicFolder() {
+  fs.copySync(paths.appPublic, paths.appBuild, {
+    dereference: true,
+    filter: file => file !== paths.appHtml
+  });
+};
